my_memes_app_backend/requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djongo==1.3.6
PyJWT==2.3.0
pymongo==3.12.1
pytz==2021.3
sqlparse==0.2.4
